+++ src/scion/integration/GaussLegendre/2/test/2.test.cpp
@@ -4,36 +4,20 @@
 using Catch::Matchers::WithinRel;
 
 // what we are testing
-#include "scion/integration/GaussLegendre.hpp"
+#include "scion/integration/GaussLegendre/2.hpp"
 
 // other includes
 
 // convenience typedefs
 using namespace njoy::scion;
 
-SCENARIO( "Gauss-Legendre quadrature rules" ) {
+SCENARIO( "Gauss-Legendre 2-point quadrature rule" ) {
 
-  GIVEN( "Gauss-Legendre 2-point integration object" ) {
+  GIVEN( "a Gauss-Legendre 2-point integration object" ) {
 
     WHEN( "integrating the [-1,1] interval" ) {
 
       integration::GaussLegendre< 2, double > integrator{};
-
-      THEN( "the integration is performed correctly" ) {
-
-        // y = f(x) = x + 1
-        auto functor = [] ( double x ) { return x + 1.; };
-
-        CHECK_THAT( 2., WithinRel( integrator( functor ) ) );
-      } // THEN
-    } // WHEN
-  } // GIVEN
-
-  GIVEN( "Gauss-Legendre 7-point integration object" ) {
-
-    WHEN( "integrating the [-1,1] interval" ) {
-
-      integration::GaussLegendre< 7, double > integrator{};
 
       THEN( "the integration is performed correctly" ) {
