+++ src/scion/integration/GaussLegendre/2/test/CMakeLists.txt
@@ -1,0 +1,1 @@
+add_cpp_test( integration.GaussLegendre.2 2.test.cpp )
