+++ src/scion/integration/GaussLegendre/7/test/CMakeLists.txt
@@ -1,0 +1,1 @@
+add_cpp_test( integration.GaussLegendre.7 7.test.cpp )
