--- src/scion/integration/GaussLegendre/test/CMakeLists.txt
@@ -1,1 +1,0 @@
-add_cpp_test( integration.GaussLegendre GaussLegendre.test.cpp )
